# openapi/swagger.yaml
openapi: 3.0.0
info:
  title: WeHOME
  version: 1.0.0
  description: Solve Accommodation Problems
  termsOfService: http://swagger.io/terms/
  contact:
    email: 19020024@vnu.edu.vn
...
servers:
  - url: http://wehome-env.eba-ujwahtza.ap-southeast-1.elasticbeanstalk.com/
  - url: http://localhost:8889
  - url: http://113.167.196.55:8889
...
paths:
  /api/user:
    get:
      tags:
        - user
      security:
        - bearerAuth: [ ]
      summary: Get all user
      operationId: user.getUsers
      parameters:
        - name: limit
          in: query
          description: number of items per page
          required: false
          schema:
            type: number
        - name: page
          in: query
          description: index of page
          required: false
          schema:
            type: number
      responses:
        '401':
          description: Access token is missing or invalid
        '500':
          description: System error
        default:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /api/user/{userId}:
    get:
      tags:
        - user
      summary: Get user by id
      operationId: user.getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid userId supplied
          content: { }
        '404':
          description: User not found
          content: { }
        '500':
          description: System error
    put:
      tags:
        - user
      summary: Update user
      operationId: user.updateUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '401':
          description: Access token is missing or invalid
        '200':
          description: OK
          content: { }
        '400':
          description: Invalid userId supplied
          content: { }
        '404':
          description: User not found
          content: { }
        '500':
          description: System error
    delete:
      tags:
        - user
      summary: Delete user
      operationId: user.deleteUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        '401':
          description: Access token is missing or invalid
        '200':
          description: OK
          content: { }
        '400':
          description: Invalid userId supplied
          content: { }
        '404':
          description: User not found
          content: { }
        '500':
          description: System error
  /api/user/create:
    post:
      tags:
        - user
      summary: Create user
      operationId: user.createUser
      requestBody:
        description: Created user object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: OK
        '500':
          description: System error
  /api/user/filter:
    post:
      tags:
        - user
      summary: Filter user
      operationId: user.filterUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: limit
          in: query
          description: number of items per page
          required: false
          schema:
            type: number
        - name: page
          in: query
          description: index of page
          required: false
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  nullable: true
                phone:
                  type: string
                  nullable: true
                email:
                  type: string
                  nullable: true
      responses:
        '401':
          description: Access token is missing or invalid
        '500':
          description: System error
        default:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /api/user/{userId}/change-password:
    post:
      tags:
        - user
      summary: Change password
      operationId: user.changePassword
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                pasword:
                  type: string
      responses:
        '401':
          description: Access token is missing or invalid
        '400':
          description: Invalid userId/password supplied
        '200':
          description: OK
        '500':
          description: System error
  /api/user/login:
    post:
      tags:
        - user
      summary: Logs user into the system
      operationId: user.login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    default: OK
                  userId:
                    type: number
                  name:
                    type: string
                  token:
                    type: string
        '401':
          description: Invalid username/password supplied
        '500':
          description: System error
  /api/user/logout:
    post:
      tags:
        - user
      security:
        - bearerAuth: [ ]
      summary: Logs out current logged in user session
      operationId: user.logout
      responses:
        '401':
          description: Access token is missing or invalid
        '500':
          description: System error
        default:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    default: OK
                  token:
                    type: string
                    default: null
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        userId:
          type: number
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
        username:
          type: string
        password:
          type: string
      example:
        user_id: 10013
        name: Trong
        phone: '19001000'
        email: abc123@gmail.com
        username: tieuphonghy01
        password: $2a$08$IeZeErqqW9ZitijjJyFi9O7LxqUt0TEliiHAsXLhWozDIg58wyIjG
        role: client
...